+++ R/bibfix.R
@@ -1,27 +1,27 @@
-#' @rdname bibstats
-#' @importFrom stringr str_replace_all str_replace fixed
-#' @importFrom bib2df bib2df df2bib
-#' @importFrom purrr map
-#' @export
-fix_bib <- function(bib_file) {
-  if (missing(bib_file)) bib_file <- find_bib()
-  bib <- suppressMessages(bib2df::bib2df(bib_file))
-  bib$PAGES <- stringr::str_replace_all(bib$PAGES,
-                                        stringr::fixed("{\\textendash}"), "--")
-  bib$AUTHOR <- purrr::map(bib$AUTHOR, function(x) {
-    x <- stringr::str_replace(x, "^", "{")
-    x <- stringr::str_replace(x, "\\{\\{", "{")
-    x <- stringr::str_replace(x, "$", "}")
-    x <- stringr::str_replace(x, "\\}\\}", "}")
-  })
-  bib$EDITOR <- purrr::map(bib$EDITOR, function(x) {
-    x <- stringr::str_replace(x, "^", "{")
-    x <- stringr::str_replace(x, "\\{\\{", "{")
-    x <- stringr::str_replace(x, "$", "}")
-    x <- stringr::str_replace(x, "\\}\\}", "}")
-  })
-  bib2df::df2bib(bib, bib_file)
-}
+# #' @rdname bibstats
+# #' @importFrom stringr str_replace_all str_replace fixed
+# #' @importFrom bib2df bib2df df2bib
+# #' @importFrom purrr map
+# #' @export
+# fix_bib <- function(bib_file) {
+#   if (missing(bib_file)) bib_file <- find_bib()
+#   bib <- suppressMessages(bib2df::bib2df(bib_file))
+#   bib$PAGES <- stringr::str_replace_all(bib$PAGES,
+#                                         stringr::fixed("{\\textendash}"), "--")
+#   bib$AUTHOR <- purrr::map(bib$AUTHOR, function(x) {
+#     x <- stringr::str_replace(x, "^", "{")
+#     x <- stringr::str_replace(x, "\\{\\{", "{")
+#     x <- stringr::str_replace(x, "$", "}")
+#     x <- stringr::str_replace(x, "\\}\\}", "}")
+#   })
+#   bib$EDITOR <- purrr::map(bib$EDITOR, function(x) {
+#     x <- stringr::str_replace(x, "^", "{")
+#     x <- stringr::str_replace(x, "\\{\\{", "{")
+#     x <- stringr::str_replace(x, "$", "}")
+#     x <- stringr::str_replace(x, "\\}\\}", "}")
+#   })
+#   bib2df::df2bib(bib, bib_file)
+# }
 
 
 get_used_bib <- function(bib_file, rmd_file) {